refactor: extract testAzureFoundryConnection to core
Move Azure Foundry connection testing logic from desktop handlers to
core package for CLI reusability. The function tests user-provided
config before saving (different from validateAzureFoundry which
validates stored config).

- Add testAzureFoundryConnection() to azure-foundry.ts
- Handles API key and Entra ID authentication
- Includes retry logic for max_completion_tokens vs max_tokens
- Removes 84 lines from handlers.ts
